+++ Plugins/Creature/Creature.hpp
@@ -140,6 +140,8 @@ private:
     ArgumentStack SetCreaturePersonalSpace      (ArgumentStack&& args);
     ArgumentStack GetHeight                     (ArgumentStack&& args);
     ArgumentStack SetHeight                     (ArgumentStack&& args);
+    ArgumentStack GetHitDistance                (ArgumentStack&& args);
+    ArgumentStack SetHitDistance                (ArgumentStack&& args);
 
     CNWSCreature *creature(ArgumentStack& args);
     std::unordered_map<uint8_t, std::unordered_map<ObjectID, int16_t>> m_RollModifier;