+++ CHANGELOG.md
@@ -23,6 +23,7 @@ https://github.com/nwnxee/unified/compare/build8193.20...HEAD
 - Creature: {Get/Set}PersonalSpace()
 - Creature: {Get/Set}CreaturePersonalSpace()
 - Creature: {Get/Set}Height()
+- Creature: {Get/Set}HitDistance()
 - Encounter: GetGeometry()
 - Encounter: SetGeometry()
 - Util: GetInstructionLimit()